countdown timer and circular progress bar synced. code cleaned up

+++ app/src/main/java/com/example/tomeastman/pomodorotimer/ActiveData.java
@@ -2,7 +2,7 @@ package com.example.tomeastman.pomodorotimer;
 
 public class ActiveData {
 
-    public static int TimeInMinutes = 1;
-    public static int TimeInSeconds = TimeInMinutes * 60;
-    public static int TimeInMilliseconds = TimeInSeconds * 1000;
+    public static long TimeInMinutes = 1;
+    public static long TimeInSeconds = TimeInMinutes * 60;
+    public static long TimeInMilliseconds = TimeInSeconds * 1000;
 }
